Split helmdeployer, add list test
* refactor to be able to inject
* check error when setting up the globalCfg for helm
